清理SOUI中x64的编译警告。

+++ SOUI/include/control/STabCtrl.h
@@ -238,9 +238,9 @@ namespace SOUI
         *
         * Describe  获取tab页面数
         */
-        size_t GetItemCount()
+        int GetItemCount()
         {
-            return m_lstPages.GetCount();
+            return (int)m_lstPages.GetCount();
         }
         /**
         * STabCtrl::GetItem
