+++ SOUI/src/control/SMCListView.cpp
@@ -882,13 +882,13 @@ void SMCListView::OnKeyDown( TCHAR nChar, UINT nRepCnt, UINT nFlags )
         {
             if(!m_lstItems.IsEmpty())
             {
-                nNewSelItem = m_lstItems.GetHead().pItem->GetItemIndex();
+                nNewSelItem = (int)m_lstItems.GetHead().pItem->GetItemIndex();
             }
         }else if(nChar == VK_NEXT || nChar == VK_END)
         {
             if(!m_lstItems.IsEmpty())
             {
-                nNewSelItem = m_lstItems.GetTail().pItem->GetItemIndex();
+                nNewSelItem = (int)m_lstItems.GetTail().pItem->GetItemIndex();
             }
         }
     }
